+++ examples/swr-site/components/video.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useEffect, useRef } from 'react'
 import { useInView } from 'react-intersection-observer'
-import 'intersection-observer'
 
 export default function Video({ src, caption, ratio }) {
   const [inViewRef, inView] = useInView({ threshold: 1 })